refactor(build-cli): Move function only used in tests to test file
